quic: separate state into separate files and other cleanups
