+++ lib/internal/quic/quic.js
@@ -5,24 +5,20 @@
 /* c8 ignore start */
 
 const {
-  ArrayBuffer,
   ArrayIsArray,
   ArrayPrototypePush,
-  DataView,
-  DataViewPrototypeGetBigInt64,
-  DataViewPrototypeGetBigUint64,
-  DataViewPrototypeGetUint8,
-  DataViewPrototypeSetUint8,
   ObjectDefineProperties,
   SymbolAsyncDispose,
   Uint8Array,
 } = primordials;
 
+// QUIC requires that Node.js be compiled with crypto support.
 const {
   assertCrypto,
 } = require('internal/util');
-const { inspect } = require('internal/util/inspect');
 assertCrypto();
+
+const { inspect } = require('internal/util/inspect');
 
 const {
   Endpoint: Endpoint_,
@@ -51,47 +47,9 @@ const {
   CLOSECONTEXT_RECEIVE_FAILURE: kCloseContextReceiveFailure,
   CLOSECONTEXT_SEND_FAILURE: kCloseContextSendFailure,
   CLOSECONTEXT_START_FAILURE: kCloseContextStartFailure,
-
-  IDX_STATE_SESSION_PATH_VALIDATION,
-  IDX_STATE_SESSION_VERSION_NEGOTIATION,
-  IDX_STATE_SESSION_DATAGRAM,
-  IDX_STATE_SESSION_SESSION_TICKET,
-  IDX_STATE_SESSION_CLOSING,
-  IDX_STATE_SESSION_GRACEFUL_CLOSE,
-  IDX_STATE_SESSION_SILENT_CLOSE,
-  IDX_STATE_SESSION_STATELESS_RESET,
-  IDX_STATE_SESSION_DESTROYED,
-  IDX_STATE_SESSION_HANDSHAKE_COMPLETED,
-  IDX_STATE_SESSION_HANDSHAKE_CONFIRMED,
-  IDX_STATE_SESSION_STREAM_OPEN_ALLOWED,
-  IDX_STATE_SESSION_PRIORITY_SUPPORTED,
-  IDX_STATE_SESSION_WRAPPED,
-  IDX_STATE_SESSION_LAST_DATAGRAM_ID,
-
-  IDX_STATE_ENDPOINT_BOUND,
-  IDX_STATE_ENDPOINT_RECEIVING,
-  IDX_STATE_ENDPOINT_LISTENING,
-  IDX_STATE_ENDPOINT_CLOSING,
-  IDX_STATE_ENDPOINT_BUSY,
-  IDX_STATE_ENDPOINT_PENDING_CALLBACKS,
-
-  IDX_STATE_STREAM_ID,
-  IDX_STATE_STREAM_FIN_SENT,
-  IDX_STATE_STREAM_FIN_RECEIVED,
-  IDX_STATE_STREAM_READ_ENDED,
-  IDX_STATE_STREAM_WRITE_ENDED,
-  IDX_STATE_STREAM_DESTROYED,
-  IDX_STATE_STREAM_PAUSED,
-  IDX_STATE_STREAM_RESET,
-  IDX_STATE_STREAM_HAS_READER,
-  IDX_STATE_STREAM_WANTS_BLOCK,
-  IDX_STATE_STREAM_WANTS_HEADERS,
-  IDX_STATE_STREAM_WANTS_RESET,
-  IDX_STATE_STREAM_WANTS_TRAILERS,
 } = internalBinding('quic');
 
 const {
-  isArrayBuffer,
   isArrayBufferView,
 } = require('util/types');
 
@@ -127,7 +85,6 @@ const {
 } = require('internal/validators');
 
 const kEmptyObject = { __proto__: null };
-const kEnumerable = { __proto__: null, enumerable: true };
 
 const {
   kBlocked,
@@ -156,6 +113,12 @@ const {
   QuicStreamStats,
   QuicSessionStats,
 } = require('internal/quic/stats');
+
+const {
+  QuicEndpointState,
+  QuicSessionState,
+  QuicStreamState,
+} = require('internal/quic/state');
 
 /**
  * @typedef {import('../socketaddress.js').SocketAddress} SocketAddress
@@ -522,156 +485,6 @@ function processTlsOptions(tls) {
   };
 }
 
-class QuicStreamState {
-  /** @type {DataView} */
-  #handle;
-
-  /**
-   * @param {ArrayBuffer} buffer
-   */
-  constructor(buffer) {
-    if (!isArrayBuffer(buffer)) {
-      throw new ERR_INVALID_ARG_TYPE('buffer', ['ArrayBuffer'], buffer);
-    }
-    this.#handle = new DataView(buffer);
-  }
-
-  /** @type {bigint} */
-  get id() {
-    return DataViewPrototypeGetBigInt64(this.#handle, IDX_STATE_STREAM_ID);
-  }
-
-  /** @type {boolean} */
-  get finSent() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_STREAM_FIN_SENT);
-  }
-
-  /** @type {boolean} */
-  get finReceived() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_STREAM_FIN_RECEIVED);
-  }
-
-  /** @type {boolean} */
-  get readEnded() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_STREAM_READ_ENDED);
-  }
-
-  /** @type {boolean} */
-  get writeEnded() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_STREAM_WRITE_ENDED);
-  }
-
-  /** @type {boolean} */
-  get destroyed() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_STREAM_DESTROYED);
-  }
-
-  /** @type {boolean} */
-  get paused() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_STREAM_PAUSED);
-  }
-
-  /** @type {boolean} */
-  get reset() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_STREAM_RESET);
-  }
-
-  /** @type {boolean} */
-  get hasReader() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_STREAM_HAS_READER);
-  }
-
-  /** @type {boolean} */
-  get wantsBlock() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_STREAM_WANTS_BLOCK);
-  }
-
-  /** @type {boolean} */
-  set wantsBlock(val) {
-    DataViewPrototypeSetUint8(this.#handle, IDX_STATE_STREAM_WANTS_BLOCK, val ? 1 : 0);
-  }
-
-  /** @type {boolean} */
-  get wantsHeaders() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_STREAM_WANTS_HEADERS);
-  }
-
-  /** @type {boolean} */
-  set wantsHeaders(val) {
-    DataViewPrototypeSetUint8(this.#handle, IDX_STATE_STREAM_WANTS_HEADERS, val ? 1 : 0);
-  }
-
-  /** @type {boolean} */
-  get wantsReset() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_STREAM_WANTS_RESET);
-  }
-
-  /** @type {boolean} */
-  set wantsReset(val) {
-    DataViewPrototypeSetUint8(this.#handle, IDX_STATE_STREAM_WANTS_RESET, val ? 1 : 0);
-  }
-
-  /** @type {boolean} */
-  get wantsTrailers() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_STREAM_WANTS_TRAILERS);
-  }
-
-  /** @type {boolean} */
-  set wantsTrailers(val) {
-    DataViewPrototypeSetUint8(this.#handle, IDX_STATE_STREAM_WANTS_TRAILERS, val ? 1 : 0);
-  }
-
-  toJSON() {
-    return {
-      __proto__: null,
-      id: `${this.id}`,
-      finSent: this.finSent,
-      finReceived: this.finReceived,
-      readEnded: this.readEnded,
-      writeEnded: this.writeEnded,
-      destroyed: this.destroyed,
-      paused: this.paused,
-      reset: this.reset,
-      hasReader: this.hasReader,
-      wantsBlock: this.wantsBlock,
-      wantsHeaders: this.wantsHeaders,
-      wantsReset: this.wantsReset,
-      wantsTrailers: this.wantsTrailers,
-    };
-  }
-
-  [kInspect](depth, options) {
-    if (depth < 0)
-      return this;
-
-    const opts = {
-      ...options,
-      depth: options.depth == null ? null : options.depth - 1,
-    };
-
-    return `StreamState ${inspect({
-      id: this.id,
-      finSent: this.finSent,
-      finReceived: this.finReceived,
-      readEnded: this.readEnded,
-      writeEnded: this.writeEnded,
-      destroyed: this.destroyed,
-      paused: this.paused,
-      reset: this.reset,
-      hasReader: this.hasReader,
-      wantsBlock: this.wantsBlock,
-      wantsHeaders: this.wantsHeaders,
-      wantsReset: this.wantsReset,
-      wantsTrailers: this.wantsTrailers,
-    }, opts)}`;
-  }
-
-  [kFinishClose]() {
-    // When the stream is destroyed, the state is no longer available.
-    this.#handle = new DataView(new ArrayBuffer(0));
-  }
-}
-
 class QuicStream {
   /** @type {object} */
   #handle;
@@ -702,7 +515,7 @@ class QuicStream {
   constructor(config, handle, session, direction) {
     validateObject(config, 'config');
     this.#stats = new QuicStreamStats(kPrivateConstructor, handle.stats);
-    this.#state = new QuicStreamState(handle.stats);
+    this.#state = new QuicStreamState(kPrivateConstructor, handle.stats);
     const {
       onblocked,
       onerror,
@@ -794,171 +607,6 @@ class QuicStream {
   }
 }
 
-class SessionState {
-  /** @type {DataView} */
-  #handle;
-
-  /**
-   * @param {ArrayBuffer} buffer
-   */
-  constructor(buffer) {
-    if (!isArrayBuffer(buffer)) {
-      throw new ERR_INVALID_ARG_TYPE('buffer', ['ArrayBuffer'], buffer);
-    }
-    this.#handle = new DataView(buffer);
-  }
-
-  /** @type {boolean} */
-  get hasPathValidationListener() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_SESSION_PATH_VALIDATION);
-  }
-
-  /** @type {boolean} */
-  set hasPathValidationListener(val) {
-    DataViewPrototypeSetUint8(this.#handle, IDX_STATE_SESSION_PATH_VALIDATION, val ? 1 : 0);
-  }
-
-  /** @type {boolean} */
-  get hasVersionNegotiationListener() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_SESSION_VERSION_NEGOTIATION);
-  }
-
-  /** @type {boolean} */
-  set hasVersionNegotiationListener(val) {
-    DataViewPrototypeSetUint8(this.#handle, IDX_STATE_SESSION_VERSION_NEGOTIATION, val ? 1 : 0);
-  }
-
-  /** @type {boolean} */
-  get hasDatagramListener() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_SESSION_DATAGRAM);
-  }
-
-  /** @type {boolean} */
-  set hasDatagramListener(val) {
-    DataViewPrototypeSetUint8(this.#handle, IDX_STATE_SESSION_DATAGRAM, val ? 1 : 0);
-  }
-
-  /** @type {boolean} */
-  get hasSessionTicketListener() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_SESSION_SESSION_TICKET);
-  }
-
-  /** @type {boolean} */
-  set hasSessionTicketListener(val) {
-    DataViewPrototypeSetUint8(this.#handle, IDX_STATE_SESSION_SESSION_TICKET, val ? 1 : 0);
-  }
-
-  /** @type {boolean} */
-  get isClosing() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_SESSION_CLOSING);
-  }
-
-  /** @type {boolean} */
-  get isGracefulClose() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_SESSION_GRACEFUL_CLOSE);
-  }
-
-  /** @type {boolean} */
-  get isSilentClose() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_SESSION_SILENT_CLOSE);
-  }
-
-  /** @type {boolean} */
-  get isStatelessReset() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_SESSION_STATELESS_RESET);
-  }
-
-  /** @type {boolean} */
-  get isDestroyed() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_SESSION_DESTROYED);
-  }
-
-  /** @type {boolean} */
-  get isHandshakeCompleted() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_SESSION_HANDSHAKE_COMPLETED);
-  }
-
-  /** @type {boolean} */
-  get isHandshakeConfirmed() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_SESSION_HANDSHAKE_CONFIRMED);
-  }
-
-  /** @type {boolean} */
-  get isStreamOpenAllowed() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_SESSION_STREAM_OPEN_ALLOWED);
-  }
-
-  /** @type {boolean} */
-  get isPrioritySupported() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_SESSION_PRIORITY_SUPPORTED);
-  }
-
-  /** @type {boolean} */
-  get isWrapped() {
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_SESSION_WRAPPED);
-  }
-
-  /** @type {bigint} */
-  get lastDatagramId() {
-    return DataViewPrototypeGetBigUint64(this.#handle, IDX_STATE_SESSION_LAST_DATAGRAM_ID);
-  }
-
-  toJSON() {
-    return {
-      __proto__: null,
-      hasPathValidationListener: this.hasPathValidationListener,
-      hasVersionNegotiationListener: this.hasVersionNegotiationListener,
-      hasDatagramListener: this.hasDatagramListener,
-      hasSessionTicketListener: this.hasSessionTicketListener,
-      isClosing: this.isClosing,
-      isGracefulClose: this.isGracefulClose,
-      isSilentClose: this.isSilentClose,
-      isStatelessReset: this.isStatelessReset,
-      isDestroyed: this.isDestroyed,
-      isHandshakeCompleted: this.isHandshakeCompleted,
-      isHandshakeConfirmed: this.isHandshakeConfirmed,
-      isStreamOpenAllowed: this.isStreamOpenAllowed,
-      isPrioritySupported: this.isPrioritySupported,
-      isWrapped: this.isWrapped,
-      lastDatagramId: `${this.lastDatagramId}`,
-    };
-  }
-
-  [kInspect](depth, options) {
-    if (depth < 0)
-      return this;
-
-    const opts = {
-      ...options,
-      depth: options.depth == null ? null : options.depth - 1,
-    };
-
-    return `SessionState ${inspect({
-      hasPathValidationListener: this.hasPathValidationListener,
-      hasVersionNegotiationListener: this.hasVersionNegotiationListener,
-      hasDatagramListener: this.hasDatagramListener,
-      hasSessionTicketListener: this.hasSessionTicketListener,
-      isClosing: this.isClosing,
-      isGracefulClose: this.isGracefulClose,
-      isSilentClose: this.isSilentClose,
-      isStatelessReset: this.isStatelessReset,
-      isDestroyed: this.isDestroyed,
-      isHandshakeCompleted: this.isHandshakeCompleted,
-      isHandshakeConfirmed: this.isHandshakeConfirmed,
-      isStreamOpenAllowed: this.isStreamOpenAllowed,
-      isPrioritySupported: this.isPrioritySupported,
-      isWrapped: this.isWrapped,
-      lastDatagramId: this.lastDatagramId,
-    }, opts)}`;
-  }
-
-  [kFinishClose]() {
-    // Snapshot the state into a new DataView since the underlying
-    // buffer will be destroyed.
-    this.#handle = new DataView(new ArrayBuffer(0));
-  }
-}
-
 class Session {
   /** @type {Endpoint} */
   #endpoint = undefined;
@@ -970,7 +618,7 @@ class Session {
   #pendingClose;
   /** @type {SocketAddress|undefined} */
   #remoteAddress = undefined;
-  /** @type {SessionState} */
+  /** @type {QuicSessionState} */
   #state;
   /** @type {QuicSessionStats} */
   #stats;
@@ -1002,7 +650,7 @@ class Session {
   constructor(config, streamConfig, handle, endpoint) {
     validateObject(config, 'config');
     this.#stats = new QuicSessionStats(kPrivateConstructor, handle.stats);
-    this.#state = new SessionState(handle.state);
+    this.#state = new QuicSessionState(kPrivateConstructor, handle.state);
     const {
       ondatagram,
       ondatagramstatus,
@@ -1052,7 +700,7 @@ class Session {
   /** @type {QuicSessionStats} */
   get stats() { return this.#stats; }
 
-  /** @type {SessionState} */
+  /** @type {QuicSessionState} */
   get state() { return this.#state; }
 
   /** @type {Endpoint} */
@@ -1297,111 +945,6 @@ class Session {
   async [SymbolAsyncDispose]() { await this.close(); }
 }
 
-class EndpointState {
-  /** @type {DataView} */
-  #handle;
-
-  /**
-   * @param {ArrayBuffer} buffer
-   */
-  constructor(buffer) {
-    if (!isArrayBuffer(buffer)) {
-      throw new ERR_INVALID_ARG_TYPE('buffer', ['ArrayBuffer'], buffer);
-    }
-    this.#handle = new DataView(buffer);
-  }
-
-  #assertNotClosed() {
-    if (this.#handle.byteLength === 0) {
-      throw new ERR_INVALID_STATE('Endpoint is closed');
-    }
-  }
-
-  /** @type {boolean} */
-  get isBound() {
-    this.#assertNotClosed();
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_ENDPOINT_BOUND);
-  }
-
-  /** @type {boolean} */
-  get isReceiving() {
-    this.#assertNotClosed();
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_ENDPOINT_RECEIVING);
-  }
-
-  /** @type {boolean} */
-  get isListening() {
-    this.#assertNotClosed();
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_ENDPOINT_LISTENING);
-  }
-
-  /** @type {boolean} */
-  get isClosing() {
-    this.#assertNotClosed();
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_ENDPOINT_CLOSING);
-  }
-
-  /** @type {boolean} */
-  get isBusy() {
-    this.#assertNotClosed();
-    return !!DataViewPrototypeGetUint8(this.#handle, IDX_STATE_ENDPOINT_BUSY);
-  }
-
-  /**
-   * The number of underlying callbacks that are pending. If the session
-   * is closing, these are the number of callbacks that the session is
-   * waiting on before it can be closed.
-   * @type {bigint}
-   */
-  get pendingCallbacks() {
-    this.#assertNotClosed();
-    return DataViewPrototypeGetBigUint64(this.#handle, IDX_STATE_ENDPOINT_PENDING_CALLBACKS);
-  }
-
-  toJSON() {
-    if (this.#handle.byteLength === 0) return {};
-    return {
-      __proto__: null,
-      isBound: this.isBound,
-      isReceiving: this.isReceiving,
-      isListening: this.isListening,
-      isClosing: this.isClosing,
-      isBusy: this.isBusy,
-      pendingCallbacks: `${this.pendingCallbacks}`,
-    };
-  }
-
-  [kInspect](depth, options) {
-    if (depth < 0)
-      return this;
-
-    if (this.#handle.byteLength === 0) {
-      return 'EndpointState { <Closed> }';
-    }
-
-    const opts = {
-      ...options,
-      depth: options.depth == null ? null : options.depth - 1,
-    };
-
-    return `EndpointState ${inspect({
-      isBound: this.isBound,
-      isReceiving: this.isReceiving,
-      isListening: this.isListening,
-      isClosing: this.isClosing,
-      isBusy: this.isBusy,
-      pendingCallbacks: this.pendingCallbacks,
-    }, opts)}`;
-  }
-
-  [kFinishClose]() {
-    // Snapshot the state into a new DataView since the underlying
-    // buffer will be destroyed.
-    if (this.#handle.byteLength === 0) return;
-    this.#handle = new DataView(new ArrayBuffer(0));
-  }
-}
-
 function validateStreamConfig(config, name = 'config') {
   validateObject(config, name);
   if (config.onerror !== undefined)
@@ -1463,7 +1006,7 @@ class Endpoint {
   #pendingError = undefined;
   /** @type {Session[]} */
   #sessions = [];
-  /** @type {EndpointState} */
+  /** @type {QuicEndpointState} */
   #state;
   /** @type {QuicEndpointStats} */
   #stats;
@@ -1553,13 +1096,13 @@ class Endpoint {
     });
     this.#handle[kOwner] = this;
     this.#stats = new QuicEndpointStats(kPrivateConstructor, this.#handle.stats);
-    this.#state = new EndpointState(this.#handle.state);
+    this.#state = new QuicEndpointState(kPrivateConstructor, this.#handle.state);
   }
 
   /** @type {QuicEndpointStats} */
   get stats() { return this.#stats; }
 
-  /** @type {EndpointState} */
+  /** @type {QuicEndpointState} */
   get state() { return this.#state; }
 
   get #isClosedOrClosing() {
@@ -1830,124 +1373,6 @@ class Endpoint {
   }
 };
 
-ObjectDefineProperties(QuicStreamStats.prototype, {
-  createdAt: kEnumerable,
-  receivedAt: kEnumerable,
-  ackedAt: kEnumerable,
-  closingAt: kEnumerable,
-  destroyedAt: kEnumerable,
-  bytesReceived: kEnumerable,
-  bytesSent: kEnumerable,
-  maxOffset: kEnumerable,
-  maxOffsetAcknowledged: kEnumerable,
-  maxOffsetReceived: kEnumerable,
-  finalSize: kEnumerable,
-});
-ObjectDefineProperties(QuicStreamState.prototype, {
-  id: kEnumerable,
-  finSent: kEnumerable,
-  finReceived: kEnumerable,
-  readEnded: kEnumerable,
-  writeEnded: kEnumerable,
-  destroyed: kEnumerable,
-  paused: kEnumerable,
-  reset: kEnumerable,
-  hasReader: kEnumerable,
-  wantsBlock: kEnumerable,
-  wantsHeaders: kEnumerable,
-  wantsReset: kEnumerable,
-  wantsTrailers: kEnumerable,
-});
-ObjectDefineProperties(QuicStream.prototype, {
-  stats: kEnumerable,
-  state: kEnumerable,
-  session: kEnumerable,
-  id: kEnumerable,
-});
-ObjectDefineProperties(QuicSessionStats.prototype, {
-  createdAt: kEnumerable,
-  closingAt: kEnumerable,
-  destroyedAt: kEnumerable,
-  handshakeCompletedAt: kEnumerable,
-  handshakeConfirmedAt: kEnumerable,
-  gracefulClosingAt: kEnumerable,
-  bytesReceived: kEnumerable,
-  bytesSent: kEnumerable,
-  bidiInStreamCount: kEnumerable,
-  bidiOutStreamCount: kEnumerable,
-  uniInStreamCount: kEnumerable,
-  uniOutStreamCount: kEnumerable,
-  lossRetransmitCount: kEnumerable,
-  maxBytesInFlights: kEnumerable,
-  bytesInFlight: kEnumerable,
-  blockCount: kEnumerable,
-  cwnd: kEnumerable,
-  latestRtt: kEnumerable,
-  minRtt: kEnumerable,
-  rttVar: kEnumerable,
-  smoothedRtt: kEnumerable,
-  ssthresh: kEnumerable,
-  datagramsReceived: kEnumerable,
-  datagramsSent: kEnumerable,
-  datagramsAcknowledged: kEnumerable,
-  datagramsLost: kEnumerable,
-});
-ObjectDefineProperties(SessionState.prototype, {
-  hasPathValidationListener: kEnumerable,
-  hasVersionNegotiationListener: kEnumerable,
-  hasDatagramListener: kEnumerable,
-  hasSessionTicketListener: kEnumerable,
-  isClosing: kEnumerable,
-  isGracefulClose: kEnumerable,
-  isSilentClose: kEnumerable,
-  isStatelessReset: kEnumerable,
-  isDestroyed: kEnumerable,
-  isHandshakeCompleted: kEnumerable,
-  isHandshakeConfirmed: kEnumerable,
-  isStreamOpenAllowed: kEnumerable,
-  isPrioritySupported: kEnumerable,
-  isWrapped: kEnumerable,
-  lastDatagramId: kEnumerable,
-});
-ObjectDefineProperties(Session.prototype, {
-  closed: kEnumerable,
-  destroyed: kEnumerable,
-  endpoint: kEnumerable,
-  path: kEnumerable,
-  state: kEnumerable,
-  stats: kEnumerable,
-});
-ObjectDefineProperties(QuicEndpointStats.prototype, {
-  createdAt: kEnumerable,
-  destroyedAt: kEnumerable,
-  bytesReceived: kEnumerable,
-  bytesSent: kEnumerable,
-  packetsReceived: kEnumerable,
-  packetsSent: kEnumerable,
-  serverSessions: kEnumerable,
-  clientSessions: kEnumerable,
-  serverBusyCount: kEnumerable,
-  retryCount: kEnumerable,
-  versionNegotiationCount: kEnumerable,
-  statelessResetCount: kEnumerable,
-  immediateCloseCount: kEnumerable,
-});
-ObjectDefineProperties(EndpointState.prototype, {
-  isBound: kEnumerable,
-  isReceiving: kEnumerable,
-  isListening: kEnumerable,
-  isClosing: kEnumerable,
-  isBusy: kEnumerable,
-  pendingCallbacks: kEnumerable,
-});
-ObjectDefineProperties(Endpoint.prototype, {
-  address: kEnumerable,
-  busy: kEnumerable,
-  closed: kEnumerable,
-  destroyed: kEnumerable,
-  state: kEnumerable,
-  stats: kEnumerable,
-});
 ObjectDefineProperties(Endpoint, {
   CC_ALGO_RENO: {
     __proto__: null,
@@ -2013,13 +1438,11 @@ module.exports = {
   Endpoint,
   Session,
   QuicStream,
-  // Exported for testing
-  kFinishClose,
-  SessionState,
+  QuicSessionState,
   QuicSessionStats,
   QuicStreamState,
   QuicStreamStats,
-  EndpointState,
+  QuicEndpointState,
   QuicEndpointStats,
 };
