+++ lib/internal/quic/stats.js
@@ -100,7 +100,7 @@ class QuicEndpointStats {
    */
   constructor(privateSymbol, buffer) {
     // We use the kPrivateConstructor symbol to restrict the ability to
-    // create new instances of EndpointStats to internal code.
+    // create new instances of QuicEndpointStats to internal code.
     if (privateSymbol !== kPrivateConstructor) {
       throw new ERR_ILLEGAL_CONSTRUCTOR();
     }
@@ -175,6 +175,10 @@ class QuicEndpointStats {
     return this.#handle[IDX_STATS_ENDPOINT_IMMEDIATE_CLOSE_COUNT];
   }
 
+  toString() {
+    return JSONStringify(this.toJSON());
+  }
+
   toJSON() {
     return {
       __proto__: null,
@@ -206,7 +210,7 @@ class QuicEndpointStats {
       depth: options.depth == null ? null : options.depth - 1,
     };
 
-    return `EndpointStats ${inspect({
+    return `QuicEndpointStats ${inspect({
       connected: this.isConnected,
       createdAt: this.createdAt,
       destroyedAt: this.destroyedAt,
@@ -225,7 +229,7 @@ class QuicEndpointStats {
   }
 
   /**
-   * True if this EndpointStats object is still connected to the underlying
+   * True if this QuicEndpointStats object is still connected to the underlying
    * Endpoint stats source. If this returns false, then the stats object is
    * no longer being updated and should be considered stale.
    * @returns {boolean}
@@ -254,7 +258,7 @@ class QuicSessionStats {
    */
   constructor(privateSynbol, buffer) {
     // We use the kPrivateConstructor symbol to restrict the ability to
-    // create new instances of SessionStats to internal code.
+    // create new instances of QuicSessionStats to internal code.
     if (privateSynbol !== kPrivateConstructor) {
       throw new ERR_ILLEGAL_CONSTRUCTOR();
     }
@@ -394,6 +398,10 @@ class QuicSessionStats {
     return this.#handle[IDX_STATS_SESSION_DATAGRAMS_LOST];
   }
 
+  toString() {
+    return JSONStringify(this.toJSON());
+  }
+
   toJSON() {
     return {
       __proto__: null,
@@ -438,7 +446,7 @@ class QuicSessionStats {
       depth: options.depth == null ? null : options.depth - 1,
     };
 
-    return `SessionStats ${inspect({
+    return `QuicSessionStats ${inspect({
       connected: this.isConnected,
       createdAt: this.createdAt,
       closingAt: this.closingAt,
@@ -470,7 +478,7 @@ class QuicSessionStats {
   }
 
   /**
-   * True if this SessionStats object is still connected to the underlying
+   * True if this QuicSessionStats object is still connected to the underlying
    * Session stats source. If this returns false, then the stats object is
    * no longer being updated and should be considered stale.
    * @returns {boolean}
