+++ test/parallel/test-quic-internal-endpoint-stats-state.js
@@ -13,10 +13,14 @@ describe('quic internal endpoint stats and state', { skip: !hasQuic }, () => {
     Endpoint,
     QuicStreamState,
     QuicStreamStats,
-    SessionState,
-    SessionStats,
-    kFinishClose,
+    QuicSessionState,
+    QuicSessionStats,
   } = require('internal/quic/quic');
+
+  const {
+    kFinishClose,
+    kPrivateConstructor,
+  } = require('internal/quic/symbols');
 
   const {
     inspect,
@@ -80,7 +84,7 @@ describe('quic internal endpoint stats and state', { skip: !hasQuic }, () => {
       stream: {},
     }, {});
     const Cons = endpoint.state.constructor;
-    throws(() => new Cons(1), {
+    throws(() => new Cons(kPrivateConstructor, 1), {
       code: 'ERR_INVALID_ARG_TYPE'
     });
   });
@@ -92,6 +96,7 @@ describe('quic internal endpoint stats and state', { skip: !hasQuic }, () => {
       stream: {},
     });
 
+    strictEqual(typeof endpoint.stats.isConnected, 'boolean');
     strictEqual(typeof endpoint.stats.createdAt, 'bigint');
     strictEqual(typeof endpoint.stats.destroyedAt, 'bigint');
     strictEqual(typeof endpoint.stats.bytesReceived, 'bigint');
@@ -107,6 +112,7 @@ describe('quic internal endpoint stats and state', { skip: !hasQuic }, () => {
     strictEqual(typeof endpoint.stats.immediateCloseCount, 'bigint');
 
     deepStrictEqual(Object.keys(endpoint.stats.toJSON()), [
+      'connected',
       'createdAt',
       'destroyedAt',
       'bytesReceived',
@@ -135,6 +141,7 @@ describe('quic internal endpoint stats and state', { skip: !hasQuic }, () => {
     }, {});
     strictEqual(typeof endpoint.stats.toJSON(), 'object');
     endpoint.stats[kFinishClose]();
+    strictEqual(endpoint.stats.isConnected, false);
     strictEqual(typeof endpoint.stats.destroyedAt, 'bigint');
     strictEqual(typeof endpoint.stats.toJSON(), 'object');
   });
@@ -146,7 +153,7 @@ describe('quic internal endpoint stats and state', { skip: !hasQuic }, () => {
       stream: {},
     }, {});
     const Cons = endpoint.stats.constructor;
-    throws(() => new Cons(1), {
+    throws(() => new Cons(kPrivateConstructor, 1), {
       code: 'ERR_INVALID_ARG_TYPE',
     });
   });
@@ -156,8 +163,8 @@ describe('quic internal endpoint stats and state', { skip: !hasQuic }, () => {
   // temporarily while the rest of the functionality is being
   // implemented.
   it('stream and session states', () => {
-    const streamState = new QuicStreamState(new ArrayBuffer(1024));
-    const sessionState = new SessionState(new ArrayBuffer(1024));
+    const streamState = new QuicStreamState(kPrivateConstructor, new ArrayBuffer(1024));
+    const sessionState = new QuicSessionState(kPrivateConstructor, new ArrayBuffer(1024));
 
     strictEqual(streamState.finSent, false);
     strictEqual(streamState.finReceived, false);
@@ -195,8 +202,8 @@ describe('quic internal endpoint stats and state', { skip: !hasQuic }, () => {
   });
 
   it('stream and session stats', () => {
-    const streamStats = new QuicStreamStats(new ArrayBuffer(1024));
-    const sessionStats = new SessionStats(new ArrayBuffer(1024));
+    const streamStats = new QuicStreamStats(kPrivateConstructor, new ArrayBuffer(1024));
+    const sessionStats = new QuicSessionStats(kPrivateConstructor, new ArrayBuffer(1024));
     strictEqual(streamStats.createdAt, undefined);
     strictEqual(streamStats.receivedAt, undefined);
     strictEqual(streamStats.ackedAt, undefined);
